+++ modules/spring-core/src/main/kotlin/com/explyt/spring/core/providers/SpringImplicitUsageProvider.kt
@@ -28,14 +28,16 @@ class SpringImplicitUsageProvider : ImplicitUsageProvider {
             if (element.isConstructor) {
                 return element.isMetaAnnotatedBy(IMPLICIT_CONSTRUCTOR_ANNOTATIONS)
                         || element.containingClass?.let {
-                            it.isMetaAnnotatedBy(IMPLICIT_CLASS_ANNOTATIONS)
-                            && it.constructors.size == 1
-                        } ?: false
+                    it.isMetaAnnotatedBy(IMPLICIT_CLASS_ANNOTATIONS) && it.constructors.size == 1
+                } ?: false
             }
             if (isDynamicPropertySource(element)) {
                 return true
             }
             return element.isMetaAnnotatedBy(IMPLICIT_METHOD_ANNOTATIONS)
+        }
+        if (element is PsiField) {
+            return element.isAnnotatedBy(MOCK_BEANS)
         }
         return false
     }
@@ -89,6 +91,12 @@ class SpringImplicitUsageProvider : ImplicitUsageProvider {
                 JavaEeClasses.RESOURCE.allFqns +
                 JavaEeClasses.POST_CONSTRUCT.allFqns +
                 JavaEeClasses.PRE_DESTROY.allFqns
+
+        private val MOCK_BEANS = listOf(
+            SpringCoreClasses.MOCK_BEAN,
+            SpringCoreClasses.SPY_BEAN
+        )
+
     }
 
 }