+++ modules/spring-core/src/test/kotlin/com/explyt/spring/core/service/java/SpringSearchServiceTest.kt
@@ -9,10 +9,10 @@ import com.explyt.spring.core.service.SpringSearchService
 import com.explyt.spring.core.service.SpringSearchServiceFacade
 import com.explyt.spring.test.ExplytJavaLightTestCase
 import com.explyt.spring.test.TestLibrary
+import com.intellij.openapi.module.ModuleUtilCore
 import com.intellij.openapi.util.registry.Registry
 import com.intellij.psi.PsiMethod
 import junit.framework.TestCase
-import org.jetbrains.kotlin.idea.util.projectStructure.getModule
 
 class SpringSearchServiceTest : ExplytJavaLightTestCase() {
     override val libraries: Array<TestLibrary> = arrayOf(TestLibrary.springBootAutoConfigure_3_1_1)
@@ -29,7 +29,7 @@ class SpringSearchServiceTest : ExplytJavaLightTestCase() {
 
     fun testImportComponent() {
         val virtualFile = myFixture.copyDirectoryToProject("service/importComponent", "")
-        val module = virtualFile.getModule(project)
+        val module = ModuleUtilCore.findModuleForFile(virtualFile, project)
         TestCase.assertNotNull(module)
         val beans = SpringSearchService.getInstance(project).getBeanPsiClassesAnnotatedByComponent(module!!)
         val beanNames = beans.mapNotNullTo(mutableSetOf()) { it.psiClass.qualifiedName }
@@ -39,7 +39,7 @@ class SpringSearchServiceTest : ExplytJavaLightTestCase() {
 
     fun testImportWithBean() {
         val virtualFile = myFixture.copyDirectoryToProject("service/importComponentWithBean", "")
-        val module = virtualFile.getModule(project)
+        val module = ModuleUtilCore.findModuleForFile(virtualFile, project)
         TestCase.assertNotNull(module)
         val beans = SpringSearchServiceFacade.getInstance(project).getAllActiveBeans(module!!)
         val beanNames = beans.filter { it.psiClass.qualifiedName?.startsWith("com.") == true }
@@ -51,7 +51,7 @@ class SpringSearchServiceTest : ExplytJavaLightTestCase() {
 
     fun testImportComplexWithComponentScan() {
         val virtualFile = myFixture.copyDirectoryToProject("service/importComplexWithComponentScan", "")
-        val module = virtualFile.getModule(project)
+        val module = ModuleUtilCore.findModuleForFile(virtualFile, project)
         TestCase.assertNotNull(module)
         val beans = SpringSearchServiceFacade.getInstance(project).getAllActiveBeans(module!!)
         val beanNames = beans.filter { it.psiClass.qualifiedName?.startsWith("com.") == true }
@@ -67,7 +67,7 @@ class SpringSearchServiceTest : ExplytJavaLightTestCase() {
 
     fun testComponentMissingBeanSearch() {
         val virtualFile = myFixture.copyDirectoryToProject("service/conditionalOnMissingBean", "")
-        val module = virtualFile.getModule(project)
+        val module = ModuleUtilCore.findModuleForFile(virtualFile, project)
         assertNotNull(module)
         val beans = SpringSearchServiceFacade.getInstance(project).getAllActiveBeans(module!!)
         val beanTestClass = beans.filter { it.name == "testClass" }