+++ modules/spring-core/testdata/java/inspection/suppresses/testConfiguration/src/MockTest.java
@@ -1,0 +1,14 @@
+import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.boot.test.mock.mockito.SpyBean;
+import org.springframework.stereotype.Component;
+
+@Component
+public class MockTest {
+    @MockBean
+    DemoService demoServiceMock;
+    @SpyBean
+    DemoService demoServiceSpy;
+}
+
+@Component
+class DemoService {}